test(e2e): add initial end-to-end test for android application
Added a testTag to the LandingScreen to correctly verify we're on the screen.
